Add interactive_repy_console.r2py
This is really only an early sketch of the docstring now.

Following the experiences of the SAS 2015 workshop, let's try to
make the REPL for Repy tighter, especially on crowded WiFis. The
idea is to provide a RepyV2 "interpreter" running inside a VM, into
which one can connect using Telnet or similar.

Therefore we don't need `seash` for every interaction with the
VM, which should make especially the early stages of code development
much easier to get done: You will see the live VM log on your screen
as your code runs (including error messages), so the typical `seash`
sequence of upload/start/list/show log will be speeded up a lot.