WIP: Add inline-dylink
This is an attempt to make a minimal dylink that doesn't require
the crazy repyhelper etc. to run, and could maybe be just a handful
lines of code to copy-pasta to your project so you don't have to
run things with `dylink.r2py myprogram.r2py args`.

Maybe that's not a good idea though --- WIP!