Simplify the `runtests` seash module
Remove the uploaddir functionality, fix up the comments and docstrings,
and document what we've learned: Don't expect an internal seash function
to raise an exception! (In our case this means that we would need to
poll our vessel to see if it has moved to state "Terminated" already
before we send the next "start" command. Yikes!)