Update tcprelay's docstring
This reflects the experience gathered with the previous implementation
that would set up a new connection to the relay's control port for
every call to the relayed listen socket's `getconnection`. As mentioned
in the new docstring and the previous commit message, this exhausts
ports quickly and is a bad idea due to state handling by the local TCP
stack.

It also adds a new design based on concurrent control connections,
one for meta-information and the other already prepared for the next
incoming remote host connection. Let's see how that turns out.